OAK-593: Segment-based MK
Use the same segment cache across file and MongoDB backends

git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1525946 13f79535-47bb-0310-9956-ffa450edef68

+++ oak-run/src/main/java/org/apache/jackrabbit/oak/fixture/OakRepositoryFixture.java
@@ -115,7 +115,7 @@ public abstract class OakRepositoryFixture implements RepositoryFixture {
     }
 
     public static RepositoryFixture getSegment(
-            final String host, final int port, final long cacheSize) {
+            final String host, final int port, final int cacheSize) {
         return new OakRepositoryFixture("Oak-Segment") {
             private SegmentStore[] stores;
             private Mongo mongo;
